94 official ver

+++ exercises/08_enums/enums2.rs
@@ -8,10 +8,11 @@ struct Point {
 
 #[derive(Debug)]
 enum Message {
-    Resize { width: u64, height: u64 },
+    // TODO: Define the different variants used below.
+    Resize {width: u32, height: u32},
     Move(Point),
     Echo(String),
-    ChangeColor(u8, u8, u8),
+    ChangeColor(u32, u32, u32),
     Quit,
 }
